:recycle: [repositories] Move `ProviderRegistry` to `registry` module
