+++ tests/unit/repositories/domain/test_providers.py
@@ -21,10 +21,10 @@ from cutty.repositories.domain.providers import constproviderfactory
 from cutty.repositories.domain.providers import LocalProvider
 from cutty.repositories.domain.providers import provide
 from cutty.repositories.domain.providers import Provider
-from cutty.repositories.domain.providers import ProviderRegistry
 from cutty.repositories.domain.providers import ProviderStore
 from cutty.repositories.domain.providers import remoteproviderfactory
 from cutty.repositories.domain.providers import Repository
+from cutty.repositories.domain.registry import ProviderRegistry
 from cutty.repositories.domain.revisions import Revision
 from cutty.repositories.domain.stores import Store
